slave_eeprom_listener: Add sample for handling notification support
Sample to use the notification support added to the i2c slave-eeprom backend.
The eeprom_listener receives the i2c-bus and i2c-address of the slave-eeprom
as an argument to register itself (by its pid) as the listener process.
After starting the eeprom_listener will load the contents of the romfile
(passed as a third argument) to the eeprom registers.
After a notification from the backend driver the eeprom_listener will backup
the content of the eeprom into the romfile.

Signed-off-by: Juan Gutierrez <juan.gutierrez@nxp.com>